Make Field 25 compatible with specification
The implementation for field 25 was assuming a Bank dependent format,
not directly compatible with the MT940 specification. It does additional
parsing by splitting the field into `bank_code`, `account_number` and
`account_currency`.

This patch introduces a new, specification compatible Class for field
25. It only provides the `account_identifier` field.

Backward compatibility on the API level is provided by falling back to
the original class when requested. It prints a deprecation warning in
this case.

The change is *not* compatible when dumping the objects as class names
and directly available fields have changed.